Add support for tag based versioning.
Versions driven by git.
If the current revision is tagged, then that's the version cut.
If not, for post-versioning, the current version is the tag plus an count of the
revs since the tag.
For pre-versioning, it's the logic that's in create-tarball.sh.
Additionally, a versioninfo file is added to the tree so that pkg-resources
can find out what the version is.

Change-Id: I5991345735fa14666305538c556b519160b02b4b