Add SKIP_GENERATE_AUTHORS option to setup.py
Updates the generate_authors function in setup.py so that
it will skip generation of the authors file if the
SKIP_GENERATE_AUTHORS env variable is set.

Authors doesn't currently take that long to generate but having
the option to explicitly skip it (regardless of whether .git is there)
seems like a good addition.

Change-Id: Ie18064d2450cecc86f18d0b10e1ce473253d1a51