Add a MAINTAINERS file
Following on from discussions at the summit, add a file which explains
the idea of oslo-core being a group of generalists and that each API
in the incubator should have a specialist maintainer.

I've made a quick stab at capturing the status of APIs where its most
obvious. I'm thinking we let people submit changes to add themselves
to the file so they're on record saying "I'll maintain this".

Change-Id: I5e67f94dc77f14849dea425dec51896fed0d07ef