Drop dependency on lockutils from oslo db code
While lockutils is a useful module, it registers a few oslo.config
options, which can possibly lead to unpleasant issues
(DuplicateOptError) when we split oslo.db from oslo-incubator into a
separate library and reuse both lockutils and oslo.db in target
projects. Unfortunately, we can't just remove usage of oslo.config
options from locktutils as it can break backward compatibility for
existing deployments.
The general approach of solving DuplicateOptError issues for oslo
libraries is not to depend on oslo.config as much as possible. So it's
actually worth to have a small code duplication here for the sake of
oslo.db as a standalone library in the future.

Change-Id: I7e3c5e1e9f7df0fec01e0d484e709cb2375025f1