Fix locking in migration tests
The migration test code included an external lock with a hard-coded
relative path that breaks when run in Jenkins.  It wasn't being hit
in Oslo because it was only locking for the postgres code path,
which isn't currently tested in Oslo.

This change moves the lock so that it applies to all db types and
removes the relative path.  For the locks to work properly it was
also necessary to add the lockutils wrapper to the Oslo tox.ini.

Closes-Bug: 1255608
Change-Id: I333b833022f208e48e3b7e82e8cfad2764be5ea0