Add IPAddress field type in object models
In object models, there are ipv4 address type and ipv6 address type.
However, for some fields in models.py, for instance, host_ip in compute node,
and a bunch of IPAddress variables in network, and address in other tables.

There is no field type corresponding to them, no matter whether it is v4 or
v6. So, add one more field type in object models to specify that type.

Change-Id: Ic53ed644f4bbc1e280b2f8cbb9b9c145c3d8a2e9