update.py fails silently if module is not found
update.py can be used to copy individual modules into OpenStack components by
using the --modules switch and providing a module name. However, if this module
is not found is oslo, the user is not informed that the copy failed. Print a
message in this case, and provide the path to the file that the script tried
to copy.

Change-Id: I6f811cd9a412d83a4d83f29c8f8398ca0503f7f4
Closes-Bug: #1274647