Run tests in parallel as much as possible
The rpc unit tests have dependencies between tests that make it
impossible to run them in parallel right now, and because that
module is deprecated it doesn't make sense to spend a lot of
effort fixing the problems.  In order to test the other modules
in parallel, we now do two test runs: one in parallel for
everything but rpc, and one serialized for rpc.

Closes-Bug: #1272035

Change-Id: I85ae00b0180bb6fdc760c97882e4e7901e9db2ef