Do not use the 'extend' method on a dict_items object
This method does not exist: it is meant to be used on a list. In Python 3,
keys() return a dict_items object, instead of a list in Python 2, so an error
happens. Fix this by converting the dict_items object  to a list before calling
extend().

This is enough to fix tests/unit/db/sqlalchemy/test_models.py in Python 3, so
we can enable this test.

Change-Id: Ic09af6c084e89e19ddc60058270890dbe052c85e