objects: Add several complex field types
Adds ListOfSetsOfIntegersField and DictOfIntegersField fields.

These three fields will be needed for NUMACPUPinningCell and
InstanceNUMACPUPinningCell object that will be introduced in
subsequent commits.

We also add the DictProxyField which will allow us to use any type that
can be converted back and forth between a string to be used as a key of
a Dict complex object field.

Change-Id: Id08ec2c436aa35a36fa7f146c6f530d28abec063
Blueprint: virt-driver-cpu-pinning